+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/plugin/Plugin.kt
@@ -55,6 +55,7 @@ import java.util.Date
  */
 @Suppress("TooManyFunctions")
 public interface Plugin :
+    DependencyResolver,
     QueryMembersListener,
     DeleteReactionListener,
     SendReactionListener,