+++ stream-chat-android-client/src/test/java/io/getstream/chat/android/client/DependencyResolverTest.kt
@@ -17,8 +17,8 @@
 package io.getstream.chat.android.client
 
 import io.getstream.chat.android.client.errorhandler.ErrorHandler
-import io.getstream.chat.android.client.plugin.DependencyResolver
 import io.getstream.chat.android.client.plugin.Plugin
+import io.getstream.chat.android.client.plugin.factory.PluginFactory
 import io.getstream.chat.android.client.setup.state.internal.MutableClientState
 import io.getstream.chat.android.core.internal.InternalStreamChatApi
 import io.getstream.chat.android.models.InitializationState
@@ -51,6 +51,12 @@ public class DependencyResolverTest {
         }
             .`should throw`(IllegalStateException::class)
             .`with message`("Plugin 'io.getstream.chat.android.client.DependencyResolverTest.PluginDependency' was not found. Did you init it within ChatClient?")
+
+        invoking {
+            client.resolveDependency<FactoryDependency, SomeDependency>()
+        }
+            .`should throw`(IllegalStateException::class)
+            .`with message`("Factory 'io.getstream.chat.android.client.DependencyResolverTest.FactoryDependency' was not found. Did you init it within ChatClient?")
     }
 
     @Test
@@ -58,13 +64,20 @@ public class DependencyResolverTest {
         val client = Fixture()
             .with(InitializationState.COMPLETE)
             .with(PluginDependency(emptyMap()))
+            .with(FactoryDependency(emptyMap()))
             .get()
 
         invoking {
             client.resolveDependency<PluginDependency, SomeDependency>()
         }
             .`should throw`(IllegalStateException::class)
-            .`with message`("Dependency 'io.getstream.chat.android.client.DependencyResolverTest.SomeDependency' was not resolved from plugin 'io.getstream.chat.android.client.DependencyResolverTest.PluginDependency'")
+            .`with message`("Dependency 'io.getstream.chat.android.client.DependencyResolverTest.SomeDependency' was not resolved by plugin 'io.getstream.chat.android.client.DependencyResolverTest.PluginDependency'")
+
+        invoking {
+            client.resolveDependency<FactoryDependency, SomeDependency>()
+        }
+            .`should throw`(IllegalStateException::class)
+            .`with message`("Dependency 'io.getstream.chat.android.client.DependencyResolverTest.SomeDependency' was not resolved by factory 'io.getstream.chat.android.client.DependencyResolverTest.FactoryDependency'")
     }
 
     /** This method use [initializationStatesArguments] as a source of arguments. */
@@ -76,6 +89,7 @@ public class DependencyResolverTest {
         val client = Fixture()
             .with(initializationState)
             .with(PluginDependency(emptyMap()))
+            .with(FactoryDependency(emptyMap()))
             .get()
 
         invoking {
@@ -83,6 +97,12 @@ public class DependencyResolverTest {
         }
             .`should throw`(IllegalStateException::class)
             .`with message`("ChatClient::connectUser() must be called before resolving any dependency")
+
+        invoking {
+            client.resolveDependency<FactoryDependency, SomeDependency>()
+        }
+            .`should throw`(IllegalStateException::class)
+            .`with message`("Dependency 'io.getstream.chat.android.client.DependencyResolverTest.SomeDependency' was not resolved by factory 'io.getstream.chat.android.client.DependencyResolverTest.FactoryDependency'")
     }
 
     @Test
@@ -91,11 +111,14 @@ public class DependencyResolverTest {
         val client = Fixture()
             .with(InitializationState.COMPLETE)
             .with(PluginDependency(mapOf(SomeDependency::class to expectedDependency)))
+            .with(FactoryDependency(mapOf(SomeDependency::class to expectedDependency)))
             .get()
 
-        val result = client.resolveDependency<PluginDependency, SomeDependency>()
+        val pResult = client.resolveDependency<PluginDependency, SomeDependency>()
+        val fResult = client.resolveDependency<FactoryDependency, SomeDependency>()
 
-        result `should be` expectedDependency
+        pResult `should be` expectedDependency
+        fResult `should be` expectedDependency
     }
 
     public companion object {
@@ -108,11 +131,16 @@ public class DependencyResolverTest {
     }
 
     private class Fixture {
-        var plugins: List<Plugin> = emptyList()
+        var plugins: MutableList<Plugin> = arrayListOf()
+        var pluginFactories: MutableList<PluginFactory> = arrayListOf()
         val mutableClientState: MutableClientState = mock()
 
         fun with(plugin: Plugin) = apply {
-            plugins = plugins + plugin
+            plugins.add(plugin)
+        }
+
+        fun with(factory: PluginFactory) = apply {
+            pluginFactories.add(factory)
         }
 
         fun with(state: InitializationState) = apply {
@@ -132,7 +160,7 @@ public class DependencyResolverTest {
             retryPolicy = mock(),
             appSettingsManager = mock(),
             chatSocket = mock(),
-            pluginFactories = mock(),
+            pluginFactories = pluginFactories,
             repositoryFactoryProvider = mock(),
             mutableClientState = mutableClientState,
             currentUserFetcher = mock(),
@@ -144,7 +172,7 @@ public class DependencyResolverTest {
 
     private class PluginDependency(
         private val classes: Map<KClass<*>, Any>,
-    ) : Plugin, DependencyResolver {
+    ) : Plugin {
         override val errorHandler: ErrorHandler? = null
         override fun onUserSet(user: User) {
             /** NO-OP */
@@ -152,6 +180,19 @@ public class DependencyResolverTest {
 
         override fun onUserDisconnected() {
             /** NO-OP */
+        }
+
+        @InternalStreamChatApi
+        override fun <T : Any> resolveDependency(klass: KClass<T>): T? =
+            classes[klass] as? T
+    }
+
+    private class FactoryDependency(
+        private val classes: Map<KClass<*>, Any>,
+    ) : PluginFactory {
+        override fun get(user: User): Plugin {
+            /** NO-OP */
+            TODO("not implemented")
         }
 
         @InternalStreamChatApi